Polish apache/dubbo#6043 : [Refactor] Extracting the "parameters" property to a new abstract Dubbo config class
